% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_info.R
\name{create_query}
\alias{create_query}
\title{create_query}
\usage{
create_query(keywords, pages)
}
\arguments{
\item{keywords}{charcter vector}

\item{pages}{integer vector}
}
\value{
queries
}
\description{
create vector of search pages
}
