% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile_info.R
\name{get_profile}
\alias{get_profile}
\title{get_profile}
\usage{
get_profile(name_id)
}
\arguments{
\item{keywords}{charcter vector}

\item{pages}{integer vector}
}
\value{
queries
}
\description{
gathers all base information scrapers by name_id
}
