% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_info.R
\name{get_users}
\alias{get_users}
\title{get_users}
\usage{
get_users(html)
}
\arguments{
\item{keywords}{charcter vector}

\item{pages}{integer vector}
}
\value{
queries
}
\description{
extract list of users from search request
}
