% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_info.R
\name{scrape_main}
\alias{scrape_main}
\title{scrape_main}
\usage{
scrape_main(keywords)
}
\arguments{
\item{keywords}{charcter vector}
}
\value{
queries
}
\description{
main method for the scraper which gathers all programms
}
