% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_info.R
\name{scrape_users}
\alias{scrape_users}
\title{get_users}
\usage{
scrape_users(x)
}
\arguments{
\item{keywords}{charcter vector}

\item{pages}{integer vector}
}
\value{
queries
}
\description{
extract list of users from search request
}
